# reservations-backend/api-gateway/swagger.yaml
openapi: 3.0.0
info:
  title: Hotel Reservation System API
  description: Microservices-based hotel reservation system API Gateway
  version: 1.0.0
  contact:
    name: API Support
    email: support@reservations.com
servers:
- url: http://localhost:3000/api
  description: Development server
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    user_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    availability_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    reservation_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    search_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    review_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    notification_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    payment_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
    report_bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
        errors:
          type: array
          items:
            type: string
...
    SendTestEmailDto:
      type: object
      properties: {}
    SendEmailDto:
      type: object
      properties: {}
...
paths:
  /auth/login:
    post:
      tags:
      - Authentication
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /auth/register:
    post:
      tags:
      - Authentication
      summary: User registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
                phoneNumber:
                  type: string
                role:
                  type: string
                  enum:
                  - CUSTOMER
                  - HOTEL_ADMIN
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /notifications/test:
    post:
      tags:
      - Notifications
      summary: Send test email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - to
              properties:
                to:
                  type: string
                  format: email
                  description: Email address to send test email
                subject:
                  type: string
                  description: Optional subject for test email
      responses:
        '200':
          description: Test email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/send-email:
    post:
      tags:
      - Notifications
      summary: Send custom email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - to
              - subject
              - text
              properties:
                to:
                  type: string
                  format: email
                subject:
                  type: string
                text:
                  type: string
                html:
                  type: string
                  description: HTML content (optional)
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/reservation-confirmation:
    post:
      tags:
      - Notifications
      summary: Send reservation confirmation email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - to
              - reservationData
              properties:
                to:
                  type: string
                  format: email
                reservationData:
                  type: object
                  properties:
                    id:
                      type: string
                    hotelName:
                      type: string
                    checkIn:
                      type: string
                    checkOut:
                      type: string
                    guests:
                      type: integer
                    totalPrice:
                      type: number
      responses:
        '200':
          description: Confirmation email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /test:
    post:
      operationId: NotificationController_sendTestEmail
      summary: Send test email
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendTestEmailDto'
      responses:
        '200':
          description: Test email sent successfully
        '400':
          description: Bad request
      tags:
      - Notifications
  /send-email:
    post:
      operationId: NotificationController_sendEmail
      summary: Send custom email
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailDto'
      responses:
        '200':
          description: Email sent successfully
        '400':
          description: Bad request
      tags:
      - Notifications
...